Added the high score screen and exit screen along with txt file. please include the paths to these screens via main menu and do amend changes accordingly
Added font and size formatting options along with background and foreground color to give a retro look to match the play screen.
Also replaced JTextArea with JTextPane for better modifying for future use.
Added scores to highscore.txt file to show that the code can accomodate high volume of scores and display as per needed.